requirements.txt
altair==4.2.0
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
PyMySQL==1.0.2
pytz==2022.1
statsmodels==0.13.2
streamlit==1.8.1
streamlit_metrics==0.1.0
treamlit_echarts==0.4.0
python-pptx==0.6.18
